fixes and improvements

+++ sector-service/src/main/java/com/example/sectorservice/controller/SectorController.java
@@ -89,11 +89,12 @@ public class SectorController {
         }
         List<Company> all = companyService.findAll();
         for(Company c: all){
+            if(!c.getSectorList().isEmpty()){
             for(Sector s: c.getSectorList()){
                 if(s.getId().equals(id)){
                     this.companyService.deleteSectorFromCompany(c.getId(),id);
                 }
-            }
+            }}
             companyService.update(c);
         }
         Sector s= this.sectorService.deleteById(id);
